+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed SimpleLogRecordProcessor.shutdown() not flushing pending exports
+
 ## [1.0.1-alpha] - 2026-04-05
 - Added a BaggageSpanProcessor that adds Baggage as SpanAttributes